Fixed some swiftlint warnings and factored out the zoom into a protocol with a default implementation.
