+++ GeoTrackKitExample/GeoTrackKitExample/AppDelegate.swift
@@ -43,4 +43,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
 }
-
