+++ GeoTrackKitExample/GeoTrackKitExampleTests/GeoTrackAnalyzerTests.swift
@@ -10,14 +10,14 @@ import XCTest
 import GeoTrackKit
 
 class GeoTrackAnalyzerTests: XCTestCase {
-    
+
     var track: GeoTrack?
-    
+
     override func setUp() {
         let reader = TrackReader(filename: "reference-track-1")
         track = reader.track
     }
-    
+
     func testPerformanceExample() {
         guard let track = track else {
             return XCTFail("No track")
@@ -27,16 +27,16 @@ class GeoTrackAnalyzerTests: XCTestCase {
             analyzer.calculate()
         }
     }
-    
+
 //    func testRewriteInOrder() {
 //        guard let track = track else {
 //            return XCTFail("No track")
 //        }
-//        
+//
 //        guard let data = try? JSONSerialization.data(withJSONObject: track.map, options: .prettyPrinted) else {
 //            return XCTFail("Failed to serialize the track into data")
 //        }
-//        
+//
 //        let path = "/tmp/out.json"
 //        do {
 //            try data.write(to: URL(fileURLWithPath: path))
@@ -44,14 +44,14 @@ class GeoTrackAnalyzerTests: XCTestCase {
 //            XCTFail(error.localizedDescription)
 //        }
 //    }
-    
+
     func testAnalyze() {
         guard let track = track else {
             return XCTFail("No track")
         }
         let analyzer = GeoTrackAnalyzer(track: track)
         analyzer.calculate()
-        
+
         guard let stats = analyzer.stats else {
             return XCTFail("Failed to get stats")
         }
